Ensure that class '.UI5TR_ControlFound' is displayed properly
As some identified controls are displayed 'inline', the corresponding
HTML element may not consume space at all, thus, making it impossible to
use any kind of background or border styles. To fix this, the additional
CSS class '.UI5TR_ControlFound_InlineFix' is applied when the computed
style of an HTML element is 'inline', ultimately setting the display
stlye of the element to 'inline-block'.

This may break some visual design of the page as we are overwriting user
CSS styles. This is basically a trade-off between UX and functionality.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>