Properly handle message IDs in Connection and injection
One main ability of the class 'Connection' is to manually register
callback functions for messages by passing corresponding message IDs to
and from the page injection. With this patch, the message IDs are
properly passed and checked such that the callback functions are
appropriately executed.

To test the (re-)enabled functionality, the controller 'TestDetails' is
now handled (in parts) by the message-handler singleton 'Connection'. As
an examplary message, 'getWindowInfo' is implemented.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>