Introduce class 'ConnectionMessages'
To keep the class 'Connection' clean, any methods with relation to
special messages to be sent are moved to the new class/module
'ConnectionMessages'.

The new class exhibits only static methods, so no obsolete objects are
instantiated and the interface is clean.

The message/action 'getWindowInfo' is moved to the new class.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>