Properly store tab ID and window ID from Connection
For proper handling during recording, the singleton class 'Connection'
needs to hold the tab and window/popup it is bound to through the
architecture. To ease storing, only the tab and window IDs are stored.

The window ID is obtained for now by performing a handshake with the
background script (message types 'handshake-get-window-id' and
'handshake-send-window-id').

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>