Remove dependency on 'RecordController' from controller 'Start'
As the 'RecordController' is to be stripped soon, the dependency on it
is removed from the controller 'Start'.

For this, the property '/targetUI5Version' of the model 'recordModel' is
removed. As it is only used for proper test creation, the data is now
retrieved from the window info that is retrieved anyway in the
controller 'TestDetails'.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>