Remove 'aggregationArray' workaround for sending TestItem
Until now, the aggregation information (inherent field 'aggregation') of
an TestItem object needed to be rewrapped into a proper Array (field
'aggregationArray') to avoid that the information is scrapped during the
sending of Port messages.

In the end, this scrapping is caused by a misbehaving merge of objects
in the function 'TestItem._getElementInformation', as the aggregation
information is to be stored in an object actually.
The wrong behavior is corrected by explicitly making the field
'aggregation' an object in the function
'TestItem._getElementInformation.getElementInformationDetails'.

Any rewrapping code (both ways) is removed accordingly.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>