Refactor retrieval of registered UI5 elements
For proper re-use, the retrieval of registered UI5 elements is added to
the class 'UI5ControlHelper' in terms of the functions
'getRegisteredElements' and 'getInitializedCore'.

The first added function is able to retrieve all registered UI5
elements, while the latter is needed as a helper function for all UI5
versions < v1.67.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>