Implement message "executeAction" using Connection and ConnectionMessages
To remove the dependency on 'Communication' from the controllers
'TestStep' and 'TestDetails', the message 'executeAction' is now
communicated via 'Connection' and 'ConnectionMessages'. (Note: The event
has been named 'execute' before!)

To retrieve the DOM node for the UI5 element on that the action is to be
performed, the static function 'UI5ControlHelper.getDomForControl' is
added for convenience.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>