Remove 'Communication'
As a means to implement a solution for issue #10, the to-be-obsolete
class 'Communication' is removed.

Note: There are still some references on this class left in the code,
but mostly in unused locations (e.g., in the mock-server
implementation). Uses in the controllers 'TestStep' and 'TestDetails'
are refactored in upcoming commits.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>