Remove model 'recordModel' from controller 'TestDetails'
The bound model 'recordModel' from the Recordcontroller instance is
removed from the controller 'TestDetails' as it is not needed in any
way. In detail, the initial storage, the binding to the
recording-indicator dialog, and an (unnecessary) direct access to the
model are removed from the controller.

In the view fragment 'RecordDialog', the reference to the 'recordModel'
is removed.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>