Remove 'Navigation'
As the model class 'Navigation' is only concerned with holding the
current state of recording, its functionality is integrated into the
clss 'RecordController'.

Note: This is only a raw movement of functionality, a refactoring to
improve the code significantly by encapsulating the functionality better
is to come in a future commit.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>