Properly reset RecordController on test creation
After displaying a saved recording, the creation of a new test was
messed up due to – seemingly – randomly appearing test items. These
appeared due to a non-proper reset of the RecordController model (used
as 'recordModel' in the view 'TestDisplay').

The mistake was to create a new model entirely instead of resetting the
existing one.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>