Properly reset TestItem cache on start
When the first step in using the extension is to display an already
recorded test, the display of test items was broken as the TestItem
cache has not been initialized in this context.

The problem is fixed by calling the TestItem-cache reset in any case on
starting the page injection.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>