Adapt element-highlight colors in page injection
Before, the element highlighting has been accomplished by setting a
special background color via a specific CSS class – but only the
background color of the selected element and not any nested ones. this
way, some highlightings could not be detected as the background color is
potentially overwritten by a nested element.

The problem is fixed by applying the hightlighting color also to nested
elements (e.g., '.UI5TR_ElementHover *'). Based on this fix, the CSS
class 'UI5TR_ControlFound_InlineFix' can be removed.

For better usability, the opacity of the background color for the CSS
class '.UI5TR_ElementHover' is removed. In the same vein, all colors are
transformed to RGB (instead if RGBa).

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>