Refactor 'RecordController.showPlayOnTestElementById'
The check whether an index is passed to the function
'RecordController.showPlayOnTestElementById' has been malfunctioning as
for the index 0, the check ('!0') has been triggered. This rendered the
start of replaying useless as all steps of a test recording have been
shown as replayable.

To fix the problem, the check is now performed explicitly against
'undefined' and 'null'.

The functionality of the function is refactored to use
'Array.prototype.forEach' instead of an 'for'-loop. To this end, the
bindings need to be updated appropriately.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>