Move all logic on executed actions to page injection
For a better adherence to the architecture, all logic regarding
exectured actions is moved to the page injection. That is, the decision
about an action's success is made in the page injection now.

There are several changes needed for this to work:
- The Connection message "executeAction" yields a result now, indicating
the success of the action execution.
- The method 'UI5ControlHelper.getDomForControl' now returns a more
reliable value. The documentation is updated accordingly.
- The function 'TestDetails._executeAction' is refactored to make direct
use of the Promise-based result of the sent Connection message.
- The function 'TestDetails.onReplaySingleStep' is refactored to handle
the new result from the executed action.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>