Run action execution promise-based in page injection
By running action executions using Promises, it is possible to gather
their results before returning.

For execution of actions, custom events are issued. The result of an
issued event are now validated using a custom event listener and then
passed to the promise-based action execution.

There is a timeout of 5 seconds on the custom event listener to collect
an issued custom event. After this timeout a tested action is reported
as errored.

The return value for assert execution is adapted accordingly.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>