Add replay-timeout setting and fix replay-type setting
For easier configuration of replays, the setting 'defaultReplayTimeout'
is added (see also commit 6f0dc3e2f2a595f909d1528202ec2dab36bdda71) to
the view 'Settings' and the underlying model 'settings'.
Note: The setting is not yet configurable from the UI, but will be in
the future.

Furthermore, the setting 'defaultReplayType' is fixed. Due to
inconsistent naming (i.e., 'replayType' vs. 'defaultReplayType'), the
dropdown in the view 'TestDetails' has not worked. This is fixed by
using the spelling 'defaultReplayType' consistently in the complete code
base.

Two further minor changes are carried out:
- The loading of stored setting values is refactored to be more concise
and stable towards future changes.
- Trailing whitespace in the touched files is removed.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>