Use replay timeout
The replay-timeout setting is now configurable using the view
'Settings', is passed on to the page injection, and properly respected
during action execution.

For better usability, the setting can be individually set in the view
'TestDetails'. To better distinguish replay timeout and replay type in
the UI, corresponding labels are added.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>