Fix execution of typing actions
The execution of typing event, especially the insertion of text into
input fields, relied on the manipulation of JQuery objects. As such
objects are no longer used, the insertion is now carried out in plain
JavaScript.

Furthermore, the corresponding function '_executeAction' in the
controller 'TestStep' is adapted to the same functionality as in the
controller 'TestDetails'.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>