Implement new replay functionality
The replay functionality is now placed in the RecordController – instead
of the controller 'TestDetails'. This way, the recording and replaying
functionality is properly encapsulated in the RecordController.

Furthermore, any component can subscribe to the RecordController model
now and bind its properties for convenient updates of both the recording
and replaying functionalities.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>