Refactor action/assert results to use proper messages
To improve compatibility with the current action and assert
functionality in the controller 'TestStep', the action/assert results
from the page injection now use proper messages that can be used in
MessagePopover controls.

for convenience and easier refactoring, the full potential of the new
messages is not yet used – but will be in the future.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>