Properly handle permission 'tabs' in RecordController
As the function 'chrome.permissions.request' returns 'undefined'
randomly if the requested permission is already granted, the requesting
behavior for the permission 'tabs' within the extensions is refactored
as follows:

- The method 'Utils._requestTabsPermission' is added to check
whether the permission is already granted and requests the permission if
not. The return value is a Promise to handle requests conveniently.

- The method 'RecordController.startReplaying' uses the new function,
which simplifies the code significantly.

- The method 'Start._getUI5Urls' uses the new function, which simplifies
the code significantly.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>