Reliably reset RecordController on mode change or cancellation
Two case for resetting the RecordController and its inherent Connection
are handled appropriately now:

1) When changing from recording to replay mode or vice-versa, the
RecordController is now reliably reset.

2) When cancelling a test or navigating back from the view
'TestDetails', the RecordController is also reset. If there is an
already connected tab, the user is asked whether to close it or not. The
Connection is cut off in any case.

While refactoring, the model property 'replayMode' of the controller
'TestDetails' is removed and substituted with the corresponding calls to
the RecordController instance.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>