Fortify execution state during replay
The property '/isExecuting' in the record model is now a "global" one
and is not attached to the single replay steps anymore. Also, the
property '/replayType' is now stored in the record model for more
reliability and better binding ability.

With these refactorings, the visibility and enabling of the
replay-next-step button in the view 'TestDetails' is more precisely
configured.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>