Remove UUID generation from RecordController
The UUIDv4 generation in the RecordController is superfluous as there is
also an implementation in 'Utils'. The implementation in the latter is
refactored to be more reliable (e.g., removal of dependency on
non-existing setting).

The call to the function in 'Utils' is changed also in the controller
'Start', applying the refactoring also there.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>