Controller 'TestStep' does not execute test steps anymore
Due to separation of concerns, the RecordController is the single point
of responsibility regarding the execution of test steps. To remove this
functionality from the controller 'TestStep', the following steps need
to be performed:

- The method 'RecordController._executeTestStep' is now public and
renamed to 'RecordController.executeTestStep'.
- The same method now takes an argument denoting the test step to be
executed. The method 'RecordController.replayNextStep' is refactored
accordingly.
- The method 'TestStep._executeAction' is removed, while the method
'_save' is refactored.

This change does not affect the test-step execution during the test-step
construction but only the one while saving. The other will be refactored
in upcoming commits.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>