Fix return page-injection values for page event 'checkAction'
When checking the preconditions for an action within the page injection,
the return type is adapted to be compatible with the event return types
'executeAssert' and 'executeAction':

- The returned messages have been cumbersome to deal with due to empty
objects in the case of success. This is fixed by leaving out the
property 'message' in case of success within
'__checkActionPreconditions'.

- The returned object had a property 'message' instead of 'messages'.
This is fixed by re-packaging the result within '__checkAction'.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>