Actually execute support assistant during replay
During replay, support-assistant runs are executed and evaluated now
using the Connection message 'runSupportAssistant'. This is also used
during test-step generation.

The following steps are needed to implement the proper execution during
replay:
- The functionality to run the support assistant now returns a result
that is analogous to the result of of the messages 'executeAction',
'executeAssert', and 'checkAction' (object with result value and
messages). The lists of issues and rules (the old return value) are
stored along with the new message-based return value. Corresponding
changes are carried out in the page injection, the RecordController, and
the view 'TestStep'.
- The controller 'TestStep' now inspects Connection-message results
directly, analogously to action checks and assert executions; incoming
messsages are collected appropriately.
- The method '_getAttributeRating' of the controller 'TestStep' is
adapted to the changes listed above. An unneeded Promise (including a
Connection message!) is removed.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>