Only execute actions when saving test step
As actions (such as press and click events) are prevented during
recording, such events need to be executed when saving and, thus,
finalizing a test step. All other types of test steps (asserts and
support-assistant runs) do not need to be carried out again as their
results are already gathered.

To ensure this behavior, the saved test step is inspected to be an
action before executing it while saving it.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>