Remove cache 'findItem' from page injection
The cache 'findItem' in the page injection, which is filled when finding
items on the page using a selector (see 'TestItem.findItemsBySelector'),
does interfere with proper selection of items in specific cases.

Consider the following case:
1) Open an UI5 page with a filter bar somewhere and start recording.
2) Select all items in the filter-bar control (e.g., all lines in a
table) and run an assert on the length of the resulting list.
3) Fill the filter bar to reduce the number of elements in the control.
4) Re-do Step 2, but with a reduced length of the resulting list.
5) Save the recorded test.
6) Replay the test and see the last recorded test step fail, i.e., the
assert on the reduced number of items to be found.

The test case described above fails as selected items are cached in the
page injection using the stringified selector for easier access (see
'TestItem._oTestGlobalCache' and 'TestItem.findItemsBySelector'): The
selector of both asserts is the same, thus, the original set of items is
returned for the second assert – causing an error when actually
executing the assert.

To fix this problem, the cache 'findItem' is removed completely as there
is no reliable substitution (e.g., by integrating window info into the
stringified cache key).

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>