Adjust column display in view 'TestDetails'
The column 'Activity' in the view 'TestDetails' has not been working for
a while. While re-activating this column, the column layout (also of the
column 'Type') is adapted to use Text controls instead of Select
controls.

To implement this, the dynamic selection of "translation" texts needs to
be performed. This enabled by providing a formatter in the controller
'TestDetails' (see '_formatTestStepDetails').

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>