Show warning on test-step display without page injection
When there is no page injection, the information on a test step is not
completely retrievable from the page (obviously), so a warning is shown.
The test step is not locked for changes then.

The handler for the route 'elementDisplay' is renamed to
'_onObjectMatchedDisplay', to reflect its purpose more closely.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>