Reliably set the execution state for all test-step execution
Since the setting of the execution state is handled in the method
'RecordController.replayNextStep', the page state (locked or unlocked)
is not reliably set while saving a test step (and executing it in the
process). To reliably set the execution state of a test step and the
page state, both are now set in the method
'RecordController.executeTestStep'.

This way, executed click events during action execution are not caught
by the listener 'document.onclick' anymore, which rendered the test-step
execution impossible.

Thanks to @adrianmarten for feedback in the debugging process.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>