Handle event 'pageDisconnected' in controller 'TestDetails'
When the connection to the page is lost, the view and controller
'TestDetails' now handles the event that is published through the event
bus. The user is asked whether to save the current status or not, before
a redirection to the route 'start' is performed.

To properly implement this functionality, the method
'TestDetails.onSave' returns the save promise now. Otherwise, the save
operation does not properly finish.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>