Reset PageListener and event listener if page is disconnected
When the page is disconnected and not reloaded, the event listeners are
preserved and, consequently, triggered if another page injection occurs
into the same tab. On the page event 'disconnect', the PageListener, the
PageCommunication, and all inherent event listeners are properly removed
and reset.

Additionally, the busy dialog for locking the page is destroyed to free
its ID for further usage (e.g., in another injection).

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>