Streamline project title throughout project
The project title has been handled and written differently throughout
the project. The spelling is changed according to the following rules:

- title case: UI5 Test Recorder,
- short title: UI5 Test Rec.,
- within prose: UI5 test recorder,
- Node.js module and web extension: ui5-testrecorder, and
- in constants: "UI5TR".

While the title and prose versions are based on the correct English
spelling, the Node.js module name is based on the repository name. The
usage in constants is short and precise.