Fix assertion-status update in view/controller 'TestStep'
With the changes from commit a699a20361f08f7cf28c1b3a99440d568edb76e1,
the preview update in the view 'TestStep' is performed erroneously: The
assertion status on the individual identified elements/items is updated
one step too late. The panel status is correct, while the individual
element messages that are shown are the one from the last update.

The individual messages are stored in the cache 'fnGetElementInfo'
within the page injection. So, to fix the problem, the list of
identified elements needs to be retrieved again after running assertions
in the controller 'TestStep'.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>