Reload existing tab before injection
View(-relative) IDs are generated in UI5 during navigation (e.g.,
'__xmlview3--*'), replay of recorded tests may lead to errors as new
tabs are opened for replaying. To circumvent this problem, when
connecting to an existing tab for recording, the tab is reloaded before
injecting the script.

In the end, the connection to an existing tab is established as follows:
1) Reload the page in the to-be-connected tab,
2) after reloading, wait some seconds and attempt content-script
injection afterwards, and
3) if injection is done, resolve the connection promise and, thus,
trigger further steps.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>