Stop recording when closing record dialog
During recording, the view 'TestDetails' displays the record dialog,
which is a BusyDialog control and indicates on-going recording. When
clicking the inherit button "Stop recording", the recording is to be
stopped and the page injection should not lock interaction with the page
anymore.

The release of the page locking has been broken by a regression
introduced by commit fe1163e90ca75c36586f6e1891ca84e5e97a53b2. This is
fixed by attaching the function 'onStopRecord' to the record dialog's
closing event again: The page is unlocked if and only if the user
pressed the button 'Stop recording' actively.

This fixes #43.