Fix on-the-fly update of assert execution and action checking
When handling a test item in the controller and view 'TestStep', the
state of the currently configured test step is not properly evaluated:
The on-the-fly visualization of an assert or action result (i.e., the
coloring of panel titles and the star rating) is not updated under all
circumstances. This issue is a regression caused by the changes in
commit a8e31a5fca6941eb3563873286d42bc6a0bef5c2.

This is fixed by calling the function '_validateSelectedItemNumber()' in
more places where the preview is updated.

This is related to issue #45.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>