Fix busy indicator not being reset on save in controller 'TestStep'
When saving a test step in the view 'TestStep', the function
'_checkAndDisplay()' issues a MessageBox where the user is able to
cancel the saving process. Until now, a cancelation is not handled
properly -- rendering the save button useless afterwards as its busy
indicator is not being reset.

This is fixed by rejecting the corresponding Promises if the user
cancels the saving process when being asked for that. The save buttons
in the view are set to *not* being busy then.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>