Refactor binding-related control attributes in 'TestStep'
To reduce the possibility of erroneous configurations in test-step
attributes regarding bindings, the various binding-related attributes
are reafactored into a single one called 'Binding'.

The control selector in the page injection now works on the fully
qualified binding path, including a prefix with the model name and the
binding contxt (e.g., "defaultModel>/todos/1/title").

This is a first step for issue #61. This patch does not adjust the code
generation along with the control selectors.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>