Re-add default values for identified items in page injection
The test items that are identified and packaged in the page injection
for analysis and display have several information on them. This holds
especially for non-existing parent items that are recursively packaged.

After the removal of the default values in commit e0e1a8c846, the
analysis and display broke due to missing values on non-existing parent
items. This regression is fixed by re-adding the default values.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>