Ensure availability of UI5 version in 'TestStep' while recording
For some code-generation strategies, the UI5 version is needed, for
example, to adjust for version-dependent changes to matchers. To make
the UI5 version available in the controller 'TestStep and, hence, code
generation (while recording), the version is stored in the code settings
when obtaining window information during preview update.

When an already recorded test step is displayed, the UI5 version is not
needed anymore since it is stored along with and inside the test-step
information.

Furthermore, the code settings have not been passed on during OPA5 code
generation. This is fixed by adding according parameters.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>